Resolved #686451
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@1196 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/smiles/SmilesGeneratorTest.java
@@ -1,9 +1,11 @@
-/*
- * $RCSfile$    $Author$    $Date$    $Revision$
+/* $RCSfile$    
+ * $Author$    
+ * $Date$    
+ * $Revision$
  * 
- * Copyright (C) 1997-2002  The Chemistry Development Kit (CDK) project
+ * Copyright (C) 1997-2003  The Chemistry Development Kit (CDK) project
  * 
- * Contact: steinbeck@ice.mpg.de, gezelter@maul.chem.nd.edu, egonw@sci.kun.nl
+ * Contact: cdk-devel@lists.sourceforge.net
  * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public License
@@ -88,13 +90,13 @@ public class SmilesGeneratorTest extends TestCase
 		   particluar carbon-only molecule until we have a proper 
 		   hydrogen count configurator
 		 */
-		int bondCount = 0;
+		double bondCount = 0;
 		Atom atom;
 		 for (int f = 0; f < mol.getAtomCount(); f++)
 		{
 			atom = mol.getAtomAt(f);
 			bondCount =  mol.getBondOrderSum(atom);
-			atom.setHydrogenCount(4 - bondCount - (int)atom.getCharge());
+			atom.setHydrogenCount(4 - (int)bondCount - (int)atom.getCharge());
 			if (standAlone) System.out.println("Hydrogen count for atom " + f + ": " + atom.getHydrogenCount());
 		}
 	}
