+++ src/org/openscience/cdk/test/structgen/deterministic/EquivalentClassesDeterministicGeneratorTest.java
@@ -1,12 +1,11 @@
-/*
- * $RCSfile$    
+/* $RCSfile$    
  * $Author$    
  * $Date$    
  * $Revision$
  * 
- * Copyright (C) 1997-2002  The Chemistry Development Kit (CDK) project
+ * Copyright (C) 1997-2003  The Chemistry Development Kit (CDK) project
  * 
- * Contact: steinbeck@ice.mpg.de, gezelter@maul.chem.nd.edu, egonw@sci.kun.nl
+ * Contact: cdk-devel@lists.sourceforge.net
  * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public License
@@ -76,13 +75,13 @@ public class EquivalentClassesDeterministicGeneratorTest extends TestCase
 		   particluar carbon-only molecule until we have a proper 
 		   hydrogen count configurator
 		 */
-		int bondCount = 0;
+		double bondCount = 0;
 		Atom atom;
 		 for (int f = 0; f < mol.getAtomCount(); f++)
 		{
 			atom = mol.getAtomAt(f);
 			bondCount =  mol.getBondOrderSum(atom);
-			atom.setHydrogenCount(4 - bondCount - (int)atom.getCharge());
+			atom.setHydrogenCount(4 - (int)bondCount - (int)atom.getCharge());
 			if (standAlone) System.out.println("Hydrogen count for atom " + f + ": " + atom.getHydrogenCount());
 		}
 	}
